-enh rename ads7138_registers.h
