+++ modules/hal_sensors/sensor_fingerposition/src/ads7138_registers.h
